+++ docs/src/components/Layout.js
@@ -1,311 +1,29 @@
-/* eslint-disable import/no-unresolved */
-import React, { useState } from 'react';
+import React from 'react';
 import {
   Row,
   Column,
-  Button,
-  Toast,
-  Tooltip
 } from '@innovaccer/design-system';
-import LeftNav from './LeftNav';
-import TableOfContent from './TableOfContent/TableOfContent';
+import Footer from './Footer/Footer';
 import Header from './Header';
-import Container from './Container';
-import ComponentsContainer from './Container/ComponentsContainer';
-import { MDXProvider } from "@mdx-js/react";
-import * as MDSComponents from '@innovaccer/design-system';
+import LeftNav from './LeftNav';
 import Meta from './Meta';
 import './css/style.css';
-import PropsTable from './PropsTable/index';
-import Rules from './Rules/Rules';
-import DOs from './Rules/DOs';
-import DONTs from './Rules/DONTs';
-import InlineMessage from './Rules/InlineMessage';
-import IconWrapper from './Rules/IconWrapper';
-import Footer from './Footer/Footer';
-import ProductLogos from './Logos/Logos';
-import ProductColors from './Colors/Colors';
-// import { useGetStorybookData } from '../util/StorybookData';
-import { ArgsTable } from './PropsTable/Table';
-import Markdown from 'markdown-to-jsx';
-import { useFrontmatter } from '../util/Frontmatter';
-import MDXHeading from './MDXHeading.js';
-import { copyMessage, copyMessageSuccess } from "../util/constants.js";
-import axios from 'axios';
-
-const useGetStorybookData = async (name) => {
-  let componentName = name;
-
-  if (!name.startsWith('components')) {
-    componentName = `components-${componentName}`
-  }
-
-
-  const data = await axios.get(`/sb/${componentName}.json`)
-    .then(({ data = {} }) => {
-      if (!Object.keys(data).length) {
-        return Promise.reject(`Could not get details for id: ${componentName}`);
-      }
-      return data;
-    })
-
-  return data;
-}
-
-function getJsxCode(name) {
-  // eslint-disable-next-line react-hooks/exhaustive-deps
-  return useGetStorybookData(name)
-    .then(componentData => {
-      const jsxCode = componentData && componentData.parameters
-        ? componentData.parameters.docs.docPage?.customCode ||
-        componentData.parameters.storySource?.source
-        : '';
-      return jsxCode;
-
-    })
-}
-
-function getPropTableData(name) {
-  // eslint-disable-next-line react-hooks/exhaustive-deps
-  return useGetStorybookData(name)
-    .then(componentData => {
-
-      const jsxCode = componentData
-        ? componentData.parameters.argTypes
-        : '';
-      return jsxCode;
-
-    })
-}
-
-
-const Preview = ({ name }) => {
-  return (
-    <div>
-      <PropsTable
-        dataProvider={() => getJsxCode(name)}
-      />
-    </div>
-  );
-};
-
-const A11yBlock = ({ name }) => {
-  const [data, setData] = React.useState([]);
-  const [loading, setLoading] = React.useState(true);
-  const [error, setError] = React.useState(null);
-
-  React.useEffect(() => {
-    useGetStorybookData(name)
-      .then((componentData) => {
-        // eslint-disable-next-line react-hooks/exhaustive-deps
-        const a11yProps = componentData && componentData.parameters.docs.docPage?.a11yProps;
-
-        setData(a11yProps);
-        setLoading(false);
-      })
-      .catch((err) => {
-        setError(err);
-        setLoading(false);
-      });
-  }, []);
-
-  if (loading) {
-    return <div>
-      Loading ...
-    </div>
-  }
-
-  if (error) {
-    return <MDSComponents.Message className="my-7" appearance="alert" title={error} description="Hold tight, we are working to get it up for you to interact with." />
-  }
-  return (
-    <div className="mb-8">
-      {data && <Markdown className="A11y-markdown">{data}</Markdown>}
-    </div>
-  );
-}
-
-const PreviewWithPropTable = ({ name }) => {
-
-  const [data, setData] = React.useState({});
-  const [loading, setLoading] = React.useState(true);
-  const [error, setError] = React.useState(null);
-
-  React.useEffect(() => {
-    getPropTableData(name)
-      .then((data) => {
-        setData(data);
-        setLoading(false);
-      })
-      .catch((err) => {
-        setError(`Could not get details for id: ${name}`);
-        setLoading(false);
-      });
-  }, []);
-
-  if (loading) {
-    return <div>
-      Loading ...
-    </div>
-  }
-
-  if (error) {
-    return <MDSComponents.Message className="my-7" appearance="alert" title={error} description="Hold tight, we are working to get it up for you to interact with." />
-  }
-
-  return (
-    <div className="overflow-x-scroll">
-      <ArgsTable rows={data} />
-    </div>
-  );
-};
-
-
-const List = ({ children, ...rest }) => {
-  return (
-    <div className='list'>
-      {children}
-    </div>
-  )
-}
-const leftMenuList = [
-  {
-    title: 'Masala Design System'
-  }
-];
+// import { useFrontmatter } from '../util/Frontmatter';
 
 const Layout = ({
-  children,
   titleType,
   pageTitle,
   pageDescription,
   pageKeywords,
   relativePagePath,
-  tabs,
-  logos,
   showMobile,
-  location,
-  ...rest
+  showAnimation,
+  is404Page,
+  children,
+  frontmatter
 }) => {
-  const is404 = children && children.key === null;
-  const [isToastActive, setIsToastActive] = useState(false);
-  const [codeCopyText, setCodeCopyText] = useState('')
-  const [toastTitle, setToastTitle] = useState('');
-  const [isTooltipActiveCode, setTooltipActiveCode] = useState(false)
-  const [tooltipName, setTooltipName] = useState(copyMessage);
-  const frontmatter = useFrontmatter(relativePagePath);
-  const refCode = React.createRef();
 
-  const copyToClipboard = (str) => {
-    if (typeof (str) === "string") {
-      navigator.clipboard.writeText(str);
-    }
-    else {
-      let codeBlock = '';
-      if (Object.keys(str).length > 0) {
-        const element = str.props.children;
-        if (Array.isArray(element) && element.length) {
-          element.map((elt) => {
-            if (typeof elt === 'object') {
-              codeBlock = codeBlock + elt.props.children;
-            } else {
-              codeBlock = codeBlock + elt;
-            }
-          });
-        } else {
-          codeBlock = str.props.children;
-        }
-      }
-      navigator.clipboard.writeText(codeBlock);
-      setCodeCopyText(str)
-    }
-  };
-
-  const Code = ({ children, ...rest }) => {
-    return (
-      <>
-        <div {...rest}>{children}</div>
-        <div
-          onMouseLeave={() => { setTooltipName(copyMessage); setTooltipActiveCode(false) }}
-          className='ml-auto'
-        >
-          <Tooltip
-            open={children === codeCopyText ? isTooltipActiveCode : false}
-            tooltip={tooltipName}
-            position="bottom"
-            appendToBody={true}
-            boundaryElement={refCode}
-          >
-            <Button
-              icon='copy'
-              className='p-0'
-              onClick={() => { copyToClipboard(children); toggleTooltip(copyMessageSuccess, "code") }}
-            />
-          </Tooltip>
-        </div>
-      </>
-    );
-  };
-
-  const toggleToast = (name) => {
-    setIsToastActive(true);
-    setToastTitle(name);
-    setTimeout(() => setIsToastActive(false), 1500);
-  }
-  const toggleTooltip = (name, type) => {
-    if (type === "code") {
-      setTooltipActiveCode(true)
-    }
-    setTooltipName(name);
-  }
-  const Logos = ({ children, logoData, ...rest }) => {
-    return (
-      <ProductLogos
-        logoData={logoData}
-        toggleToast={toggleToast}
-      />
-    );
-  };
-
-  const Rectangle = ({ name, ...rest }) => {
-    return (
-      <div className='rectangle'>{name}</div>
-    );
-  };
-
-  const Colors = ({ children, colorData, ...rest }) => {
-    return (
-      <ProductColors
-        colorData={colorData}
-        toggleToast={toggleToast}
-      />
-    );
-  };
-  const DSComponents = {
-    ...MDSComponents,
-    pre: Code,
-    Preview: Preview,
-    PreviewWithPropTable: PreviewWithPropTable,
-    A11yBlock: A11yBlock,
-    Rules,
-    DOs,
-    DONTs,
-    InlineMessage,
-    IconWrapper,
-    h1: (props) => <MDXHeading size='xxl' headingInfo={props} />,
-    h2: (props) => <MDXHeading size='xl' headingInfo={props} />,
-    h3: (props) => <MDXHeading size='l' headingInfo={props} />,
-    h4: (props) => <MDXHeading size="m" headingInfo={props} />,
-    h5: (props) => <MDXHeading size="s" headingInfo={props} />,
-    ul: List,
-    Logos: (props) => <Logos {...props} />,
-    Rectangle: (props) => <Rectangle {...props} />,
-    Colors: (props) => <Colors {...props} />,
-  };
-  const showAnimation = () => {
-    if (location.state?.animation === false) return false;
-    return true;
-  }
+  // const newFrontmatter = useFrontmatter(relativePagePath);
 
   return (
     <>
@@ -318,75 +36,23 @@ const Layout = ({
         relativePagePath={relativePagePath}
       />
       <Header
-        leftMenuList={leftMenuList}
         relativePagePath={relativePagePath}
       />
-      <Row style={{ height: 'calc(100vh - 48px)' }} ref={refCode}>
+      <Row style={{ height: 'calc(100vh - 48px)' }}>
         <LeftNav
-          is404Page={is404}
+          is404Page={is404Page}
           relativePagePath={relativePagePath}
           pageTitle={pageTitle}
           showMobile={showMobile}
           frontmatter={frontmatter}
         />
-        <Column className={`${showAnimation() ? "page-animation" : ''} page-scroll h-100`} id="main-container">
-          <Row className='justify-content-center'>
-            <Column className="px-12 py-8 min-vh-100 inner-left-container" size={9}>
-              {!relativePagePath.includes('components') && (
-                <Container
-                  pageTitle={pageTitle}
-                  relativePagePath={relativePagePath}
-                  tabs={tabs}
-                  pageDescription={pageDescription}
-                  logos={logos}
-                  frontmatter={frontmatter}
-                >
-                  <MDXProvider components={DSComponents}>
-                    {children}
-                  </MDXProvider>
-                </Container>
-              )}
-              {relativePagePath.includes('components') && (
-                <ComponentsContainer
-                  pageTitle={pageTitle}
-                  relativePagePath={relativePagePath}
-                  tabs={tabs}
-                  pageDescription={pageDescription}
-                  frontmatter={frontmatter}
-                >
-                  <MDXProvider components={DSComponents}>
-                    {children}
-                  </MDXProvider>
-                </ComponentsContainer>
-              )}
-            </Column>
-
-            <Column
-              size={3}
-              className="pb-6 in-page-nav position-sticky scroll-y"
-              style={{ height: 'calc(100vh - 48px)' }}
-            >
-              <TableOfContent
-                is404Page={is404}
-                relativePagePath={relativePagePath}
-                pageTitle={pageTitle}
-                location={rest.location}
-              />
-            </Column>
-          </Row>
-          {isToastActive && (
-            <Toast
-              appearance='success'
-              title={toastTitle}
-              className='position-fixed ml-5 toast'
-              onClose={() => setIsToastActive(false)}
-            />
-          )}
+        <Column className={`${showAnimation ? "page-animation" : ''} page-scroll h-100`} id="main-container">
+          {children}
           <Footer relativePagePath={relativePagePath} />
         </Column>
       </Row>
     </>
-  );
-};
+  )
+}
 
 export default Layout;
